fix: found the cause of the mobile bug
