+++ src/stories/Elevation.stories.tsx
@@ -7,6 +7,8 @@ import { Text } from '../ui/Text';
 import { useColorMode } from '../ui/hooks/useColorMode';
 import { theme } from '../ui/theme/theme';
 
+const customShadows = ['elevation1', 'elevation2', 'elevation3'];
+
 const ElevationDemo = () => {
   const { colorMode } = useColorMode();
 
@@ -16,42 +18,74 @@ const ElevationDemo = () => {
         <Text fontSize="2xl" fontWeight="bold">
           Elevation Showcase
         </Text>
-        {Object.entries(theme.semanticTokens.shadows).map(([elevationName, elevationValue]) => (
-          <Flex key={elevationName} align="center" gap={4}>
-            <Box
-              width="200px"
-              height="100px"
-              bg={colorMode === 'light' ? 'white' : 'gray.800'}
-              boxShadow={(elevationValue as any).default}
-              borderRadius="md"
-              display="flex"
-              alignItems="center"
-              justifyContent="center"
-            >
-              <Text color="black">{elevationName}</Text>
-            </Box>
-            <Text>{(elevationValue as any)[colorMode === 'light' ? 'default' : '_dark']}</Text>
-          </Flex>
-        ))}
+        <Text fontSize="2xl" fontWeight="bold">
+          Chakra UI Default Elevations
+        </Text>
+        {Object.entries(theme.shadows).map(([elevationName, elevationValue]) => {
+          if (customShadows.includes(elevationName)) return null;
+          return (
+            <Flex key={elevationName} align="center" gap={4}>
+              <Box
+                width="200px"
+                height="100px"
+                bg={colorMode === 'light' ? 'white' : 'gray.800'}
+                boxShadow={elevationValue as string}
+                borderRadius="md"
+                display="flex"
+                alignItems="center"
+                justifyContent="center"
+              >
+                <Text color="black">{elevationName}</Text>
+              </Box>
+              <Text>{elevationValue as string}</Text>
+            </Flex>
+          );
+        })}
+        <Text fontSize="2xl" fontWeight="bold">
+          Custom Elevations
+        </Text>
+        {Object.entries(theme.shadows).map(([elevationName, elevationValue]) => {
+          if (!customShadows.includes(elevationName)) return null;
+          return (
+            <Flex key={elevationName} align="center" gap={4}>
+              <Box
+                width="200px"
+                height="100px"
+                bg={colorMode === 'light' ? 'white' : 'gray.800'}
+                boxShadow={(elevationValue as any).default}
+                borderRadius="md"
+                display="flex"
+                alignItems="center"
+                justifyContent="center"
+              >
+                <Text color="black">{elevationName}</Text>
+              </Box>
+              <Text>{(elevationValue as any)[colorMode === 'light' ? 'default' : '_dark']}</Text>
+            </Flex>
+          );
+        })}
       </Flex>
       <Flex direction="column" gap={8} bg="#211F1F" p={4}>
-        {Object.entries(theme.semanticTokens.shadows).map(([elevationName, elevationValue]) => (
-          <Flex key={elevationName} align="center" gap={4}>
-            <Box
-              width="200px"
-              height="100px"
-              bg={'#211F1F'}
-              boxShadow={(elevationValue as any)._dark}
-              borderRadius="md"
-              display="flex"
-              alignItems="center"
-              justifyContent="center"
-            >
-              <Text color="white">{elevationName}</Text>
-            </Box>
-            <Text color="white">{(elevationValue as any)._dark}</Text>
-          </Flex>
-        ))}
+        {Object.entries(theme.shadows).map(([elevationName, elevationValue]) => {
+          if (!customShadows.includes(elevationName)) return null;
+          return (
+            <Flex key={elevationName} align="center" gap={4}>
+              <Box
+                width="200px"
+                height="100px"
+                bg={'#211F1F'}
+                boxShadow={(elevationValue as any)._dark}
+                borderRadius="md"
+                display="flex"
+                alignItems="center"
+                justifyContent="center"
+              >
+                <Text color="white">{elevationName}</Text>
+              </Box>
+              <Text color="white">{(elevationValue as any)._dark}</Text>
+            </Flex>
+          );
+        })}
       </Flex>
     </Stack>
   );
